desktop: replace C shims and backlight core with native V
Remove all handwritten C headers and C tests under desktop/: port the
brightness client, battery client/cache and general POSIX shim to V.
Use native BacklightState/BacklightResult and a DeviceIO interface for
production and test doubles; keep libc declarations only in platform.c.v.
Preserve Display/Battery controls and the taskbar battery indicator.
Restore original terminal descriptor flags as well as termios on exit.

Port the experimental DCP backlight core and tests to allocation-free V,
with inline resource state, unchanged calibration/wire bytes, strict
parsing, bounded buffers, unique completion tokens and recovery behavior.
Remove the custom C backlight ABI and update source staging/test runners.
Fix a u64 errno argument in the existing SMC adapter found by codegen.

Validation: 30 native V test groups / 115497 assertions pass with GCC
and Clang ASan/UBSan (leak detection disabled). Full native desktop
production build passes; complete ARM64 kernel V-to-C generation passes.
Uploaded Git tree matches the tested local tree byte-for-byte.

Scope is desktop and backlight; unrelated kernel/third-party C remains.
The real DCP backend is still missing. No hardware brightness support is
claimed, and a linked Vinix image / M1 hardware run remains unvalidated.

+++ kernel/modules/apple/smc/smc.v
@@ -225,7 +225,7 @@ fn (mut this Battery) read(handle voidptr, buf voidptr, loc u64, count u64) ?i64
 	mut sample := this.snapshots[key] or {
 		percent := this.sample
 		if percent < 0 {
-			errno.set(if percent == -4 { errno.enodev } else { errno.eio })
+			errno.set(u64(if percent == -4 { errno.enodev } else { errno.eio }))
 			return none
 		}
 		if C.vinix_smc_counter() - this.sampled_at >= 2 * this.frequency {
